refactor(shields): extract timer/process helpers into timer-control module
